Сделал 4.7 декораторы; привёл всё файлы к общему стилю наименования
