--- Practice/Lazarev/HW2.1.py
@@ -1,12 +1,0 @@
-import math
-
-
-def square(r):
-    return math.pi * r ** 2
-
-
-radius = input("Введите радиус: ")
-
-result = square(int(radius))
-
-print(f"Площадь круга: {result}")
