+++ Practice/Lazarev/HW2.1Площадь_круга.py
@@ -1,0 +1,12 @@
+import math
+
+
+def square(r):
+    return math.pi * r ** 2
+
+
+radius = input("Введите радиус: ")
+
+result = square(int(radius))
+
+print(f"Площадь круга: {result}")
