--- Practice/Lazarev/HW2.2.py
@@ -1,6 +1,0 @@
-start = input("Топлива было: ")
-end = input("Топлива осталось: ")
-distance = input("Расстояние: ")
-diff = int(start) - int(end)
-result = diff / int(distance)
-print(f"Расход бензина: {result}")
