+++ Practice/Lazarev/HW2.2Расход_бензина.py
@@ -1,0 +1,6 @@
+start = input("Топлива было: ")
+end = input("Топлива осталось: ")
+distance = input("Расстояние: ")
+diff = int(start) - int(end)
+result = diff / int(distance)
+print(f"Расход бензина: {result}")
