+++ Practice/Lazarev/HW2.3pep8.py
