+++ Practice/Lazarev/HW3.1Расчёт_периметра.py
