+++ Practice/Lazarev/HW3.2Расчёт_скорости.py
