+++ Practice/Lazarev/HW3.3.1Шифратор.py
