+++ Practice/Lazarev/HW3.3.2Дешифратор.py
