+++ Practice/Lazarev/HW3.4Замена_буквы.py
