+++ Practice/Lazarev/HW3.5Палиндром.py
