+++ Practice/Lazarev/HW4.1FizzBuzz.py
