+++ Practice/Lazarev/HW4.2Таблица_цифра_равна.py
