+++ Practice/Lazarev/HW4.3Число_из_цифр.py
