+++ Practice/Lazarev/HW4.4Спорщик.py
