+++ Practice/Lazarev/HW4.6.1Большее_число_из_введёных.py
