+++ Practice/Lazarev/HW4.6.2Большее_по_месту_вызова.py
