+++ Practice/Lazarev/HW4.7=====декоратор===== .py
@@ -1,0 +1,12 @@
+def my_decorator(func):
+
+    def equals(x):
+        return '===========' + func(x) + '==========='
+    return equals
+
+@my_decorator
+def anyfunc(y):
+    return (y)
+
+print(anyfunc('привет'))
+
