+++ Practice/Lazarev/HW4.8Камень_ножницы_бумага.py
